+++ src/main/kotlin/step1/Person.kt
@@ -1,4 +1,4 @@
-
+package step1
 
 data class Person(
     private val _name: String,
