+++ src/test/kotlin/step1/PersonAnnotationTest.kt
@@ -1,3 +1,5 @@
+package step1
+
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
