+++ src/test/kotlin/step1/PersonFeatureTest.kt
@@ -1,10 +1,12 @@
-import io.kotest.core.spec.style.FreeSpec
+package step1
+
+import io.kotest.core.spec.style.FeatureSpec
 import io.kotest.matchers.shouldBe
 
-class PersonFreeTest : FreeSpec({
+class PersonFeatureTest : FeatureSpec({
 
-    "이름 있는 인자" - {
-        "이름 있는 경우" {
+    feature("이름 있는 인자") {
+        scenario("이름 있는 경우") {
             val perple = listOf(
                 Person("김종인", 33, "종인막"),
                 Person("김종인", 33, _nickname = "종인막"),
@@ -17,23 +19,25 @@ class PersonFreeTest : FreeSpec({
         }
     }
 
-    "널 타입" - {
+    feature("널 타입") {
         val person = Person("김종인", 33, null)
         person.name shouldBe "김종인"
         person.age shouldBe 33
         person.nickname shouldBe null
     }
 
-    "기본 인자" - {
+    xfeature("실행 안될 피쳐") {
         val person = Person("김종인", 33)
         person.name shouldBe "김종인"
         person.age shouldBe 33
         person.nickname shouldBe "종인막"
     }
 
-    "데이터 클래스" - {
-        val person1 = Person("김종인", 33)
-        val person2 = Person("김종인", 33)
-        person1 shouldBe person2
+    xfeature("데이터 클래스") {
+        scenario("실행 안될 시나리오") {
+            val person1 = Person("김종인", 33)
+            val person2 = Person("김종인", 33)
+            person1 shouldBe person2
+        }
     }
 })
