+++ src/test/kotlin/step1/PersonStringTest.kt
@@ -1,3 +1,5 @@
+package step1
+
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
