+++ src/test/kotlin/step1/PersonWordTest.kt
@@ -1,3 +1,5 @@
+package step1
+
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.matchers.shouldBe
