+++ src/test/kotlin/step1/TestFunction.kt
@@ -1,3 +1,5 @@
+package step1
+
 import io.kotest.matchers.shouldBe
 
 fun testFunction() {
