+++ src/test/kotlin/step2/CalcuratorTest.kt
@@ -1,3 +1,5 @@
+package step2
+
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.datatest.withData
